+++ Reqnroll.Generator/UnitTestProvider/IUnitTestGeneratorProvider.cs
@@ -28,7 +28,5 @@ namespace Reqnroll.Generator.UnitTestProvider
         void SetTestMethodAsRow(TestClassGenerationContext generationContext, CodeMemberMethod testMethod, string scenarioTitle, string exampleSetName, string variantName, IEnumerable<KeyValuePair<string, string>> arguments);
 
         void MarkCodeMethodInvokeExpressionAsAwait(CodeMethodInvokeExpression expression);
-
-        CodeExpression GetTestWorkerIdExpression();
     }
 }