+++ Reqnroll.Generator/UnitTestProvider/NUnit3TestGeneratorProvider.cs
@@ -154,8 +154,5 @@ namespace Reqnroll.Generator.UnitTestProvider
         {
             CodeDomHelper.MarkCodeMethodInvokeExpressionAsAwait(expression);
         }
-
-        public CodeExpression GetTestWorkerIdExpression() 
-            => new CodePropertyReferenceExpression(GetTestContextExpression(), TESTCONTEXT_WORKERID_PROPERTY);
     }
 }